refactor(route): refactor route hoc & structure
- remove required from query params validation in order to use js built-in default params
- change route structure
- make components factory
- optimize and restrict types
