+++ src/route/index.ts
@@ -1,0 +1,1 @@
+export { Route } from './hoc'
